pyproject.toml
[tool.poetry]
name = "django-rest-framework2"
version = "0.1.0"
description = ""
authors = ["Kk <iankk0806@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
